pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "analyze-this"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "authlib>=1.6.8",
    "fastapi>=0.129.0",
    "firebase-admin>=7.1.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.4",
    "greenlet>=3.3.1",
    "gunicorn>=25.1.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",

    "openai>=2.21.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-httpx>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[project.optional-dependencies]
mcp-server = [
    "mcp>=1.26.0",
]

[tool.setuptools]
py-modules = []

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
